requirements.txt
PyQt6==6.5.0
python-dotenv==1.0.0
openai==1.1.1
google-generativeai==0.3.1
anthropic==0.5.0
PyPDF2==3.0.1
PyMuPDFb==1.24.6
python-docx==0.8.11
Pillow==9.5.0
reportlab==3.6.12
pdfplumber==0.11.1
openpyxl==3.1.2
python-pptx==0.6.21
beautifulsoup4==4.12.2
markdown2==2.4.10
fitz==0.0.1.dev2